requirements.txt
groq>=0.9.0
langchain>=0.2.14
langchain-groq>=0.1.9
langchain-community>=0.2.14
langchain-huggingface>=0.0.3
langchain-text-splitters>=0.2.4
langchain-core>=0.2.38
langgraph>=0.2.28
sentence-transformers>=3.0.1
faiss-cpu>=1.8.0
python-dotenv>=1.0.1